requirements.txt
streamlit
pillow
ultralytics
onnxruntime
